project/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
nltk>=3.6
matplotlib>=3.4.0
umap-learn>=0.5.0
flask>=2.0.0